refactor(admin-ui-react): improve file colocation
