+++ packages/admin-ui-react/src/__stories__/jsx.stories.tsx
@@ -8,8 +8,8 @@ import React, {
 import type { Meta } from '@storybook/react'
 import { Button as ReakitButton, Separator, Role } from 'reakit'
 
-import { createSystem } from '../createSystem'
-import { jsx, tag } from '../index'
+import { createSystem } from '../system'
+import { jsx, tag } from '../system-deprecated'
 
 const [SystemProvider] = createSystem({
   key: 'admin-ui-react',
