+++ packages/admin-ui-react/src/__stories__/system.stories.tsx
@@ -13,8 +13,12 @@ import {
   MenuButton as ReakitMenuButton,
 } from 'reakit'
 
-import { createSystem } from '../createSystem'
-import { createComponent, createHook, useElement } from '../system'
+import {
+  createSystem,
+  createComponent,
+  createHook,
+  useElement,
+} from '../system'
 
 const [SystemProvider] = createSystem({
   key: 'admin-ui-react',
