+++ packages/admin-ui-react/src/__tests__/setup.ts
@@ -4,7 +4,7 @@ import { render as baseRender } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 import serializer, { matchers } from 'jest-emotion'
 import { toHaveNoViolations } from 'jest-axe'
-import { createSystem } from '../createSystem'
+import { createSystem } from '../system'
 
 const [ThemeProvider] = createSystem({
   key: 'vtex-admin-ui',
