+++ packages/admin-ui-react/src/index.ts
@@ -1,0 +1,3 @@
+export * from './system-deprecated'
+export * from './hooks'
+export * from './system'
