--- packages/admin-ui-react/src/index.tsx
@@ -1,13 +1,0 @@
-export * from './tag'
-export * from './domElements'
-export * from './jsx'
-export * from './types'
-export * from './util'
-export * from './symbols'
-export * from './hooks/useStylesheet'
-export * from './hooks/useMediaQuery'
-export * from './hooks/useResponsiveValue'
-export * from './createSystem'
-export * from './context'
-export { IconContainer } from './createIcons'
-export * from './system'
