+++ packages/admin-ui-react/typings/vtex.phosphor.d.ts
@@ -1,11 +1,11 @@
 import type { IconOptions as PhosphorOptions } from '@vtex/phosphor-icons'
 
-import type { CsxCall } from '../src/types'
+import type { ComponentStyleProps } from '../src/system'
 
 type IconSize = 'regular' | 'small'
 
 declare module '@vtex/phosphor-icons' {
-  export interface IconOptions extends CsxCall {
+  export interface IconOptions extends ComponentStyleProps {
     title?: string
     size?: IconSize
   }
